clean/pyproject.toml
[project]
name = "vlm-sae"
version = "0.1.0"
description = "Multimodal SAE alignment — clean reproducible suite"
requires-python = ">=3.10"

dependencies = [
  "numpy>=1.26",
  "torch>=2.6.0",
  "transformers>=4.50",
  "open_clip_torch>=2.26",
  "datasets>=2.20",
  "hf_transfer>=0.1.6",
  "tqdm>=4.66",
  "scipy>=1.13",
  "matplotlib>=3.9",
  "pyyaml>=6.0",
  "Pillow>=10.0",
  "einops>=0.7",
  "safetensors>=0.4",
]

[project.optional-dependencies]
test = ["pytest>=8.0"]

[tool.setuptools]
packages = ["clean", "clean.src"]

[tool.pyright]
extraPaths = [".."]
